Restructure routes: landing at /, chat at /app
BREAKING CHANGE: Route structure has been reorganized for better UX

Before:
- / → Chat interface (no context for new visitors)
- /landing → Marketing page (hidden from visitors)

After:
- / → Landing page with value proposition and features
- /app → Chat interface

Changes:
- Move landing page from /app/landing to /app (root route)
- Move chat interface from /app to /app/app
- Update primary CTA to link to /app
- Remove /app/landing directory

This ensures new visitors see the landing page first with clear
value proposition before accessing the chat interface.
