+++ app/app/page.tsx
@@ -1,0 +1,5 @@
+import App from "../App";
+
+export default function ChatPage() {
+  return <App />;
+}
